requirements.txt
datetime
discord.py
nest-asyncio
pandas
pandas_market_calendars
plotly
polygon-api-client
python-dotenv
pytz
requests